+++ test/td_bg_web/controllers/business_concept_version_search_controller_test.exs
@@ -1,7 +1,6 @@
 defmodule TdBgWeb.BusinessConceptVersionSearchControllerTest do
   use TdBgWeb.ConnCase
 
-  import ExUnit.CaptureLog
   import Mox
 
   alias TdCore.Search.IndexWorkerMock
@@ -279,22 +278,15 @@ defmodule TdBgWeb.BusinessConceptVersionSearchControllerTest do
         end
       )
 
-      log =
-        capture_log(fn ->
-          assert %{
-                   "data" => [
-                     %{"content" => %{"Field1" => "First field", "Field2" => "Second field"}}
-                   ]
-                 } =
-                   conn
-                   |> Plug.Conn.assign(:locale, nil)
-                   |> post(Routes.business_concept_version_search_path(conn, :search), %{})
-                   |> json_response(:ok)
-        end)
-
-      assert log =~ """
-             Language is not defined in the business_concept_version_search_view
-             """
+      assert %{
+               "data" => [
+                 %{"content" => %{"Field1" => "First field", "Field2" => "Second field"}}
+               ]
+             } =
+               conn
+               |> Plug.Conn.assign(:locale, nil)
+               |> post(Routes.business_concept_version_search_path(conn, :search), %{})
+               |> json_response(:ok)
     end
 
     @tag authentication: [role: "admin"]